fixed docs
